fix(gateway): repair benchmark route publish blockers
