+++ packages/gateway/src/app-route-registrars-benchmark.ts
@@ -1,5 +1,5 @@
 import { createBenchmarkFixtureRoutes } from "./routes/benchmark-fixtures.js";
-import type { AppRouteContext } from "./app-route-types.js";
+import type { AppRouteContext } from "./app-route-support.js";
 
 export function registerBenchmarkFixtureRoutes(context: AppRouteContext): void {
   context.app.route(
