+++ packages/gateway/src/app-route-support.ts
@@ -1,3 +1,5 @@
+import { Hono } from "hono";
+import type { AppOptions } from "./app.js";
 import type { GatewayContainer } from "./container.js";
 import { AuthProfileDal } from "./modules/models/auth-profile-dal.js";
 import { ConversationProviderPinDal } from "./modules/models/conversation-pin-dal.js";
@@ -13,7 +15,37 @@ import { loadAllPlaybooks } from "./modules/playbook/loader.js";
 import { PlaybookRunner } from "./modules/playbook/runner.js";
 import { LocationService } from "./modules/location/service.js";
 import { WsEventDal } from "./modules/ws-event/dal.js";
-import type { AppRouteContext } from "./app-route-types.js";
+
+export interface AppRouteDependencies {
+  authProfileDal: AuthProfileDal;
+  pinDal: ConversationProviderPinDal;
+  configuredModelPresetDal: ConfiguredModelPresetDal;
+  executionProfileModelAssignmentDal: ExecutionProfileModelAssignmentDal;
+  routingConfigDal: RoutingConfigDal;
+  channelThreadDal: ChannelThreadDal;
+  wsEventDal: WsEventDal;
+  desktopEnvironmentDal: DesktopEnvironmentDal;
+  desktopEnvironmentHostDal: DesktopEnvironmentHostDal;
+}
+
+export interface AppRouteContext {
+  app: Hono;
+  container: GatewayContainer;
+  opts: AppOptions;
+  runtime: {
+    version: string;
+    instanceId: string;
+    role: string;
+    otelEnabled: boolean;
+    desktopTakeoverAdvertiseOrigin?: string;
+  };
+  isLocalOnly: boolean;
+  channelPipelineEnabled: boolean;
+  wsMaxBufferedBytes?: number;
+  workflowRunner: AppOptions["workflowRunner"];
+  secretProviderForTenant: AppOptions["secretProviderForTenant"];
+  routeDeps: AppRouteDependencies;
+}
 
 export function createAppRouteDependencies(container: GatewayContainer) {
   return {
